% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_chem_alerts.R
\name{med_chem_alerts}
\alias{med_chem_alerts}
\title{Look up Medicinal Chemistry Alerts}
\usage{
med_chem_alerts(smiles, ...)
}
\arguments{
\item{smiles}{a SMILES string}

\item{api_key}{an API_KEY}
}
\value{
tibble::tibble with columns ['smiles', 'alerts']
  where smiles is the input smiles and alerts is nested data.frame
}
\description{
See the blog post: Manifold’s new alerts for nuisance compounds in cellular assays
https://discuss.postera.ai/t/manifolds-new-alerts-for-nuisance-compounds-in-cellular-assays/2853
}
\details{
For additional optional parameters, see posetra_REST()
}
