% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substance_search.R
\name{search_exact}
\alias{search_exact}
\title{Search for  substance exactly}
\usage{
search_exact(smiles, ...)
}
\arguments{
\item{smiles}{a SMILES string}

\item{api_key}{an API_KEY}
}
\description{
For additional optional parameters, see posetra_REST()
}
