% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substance_search.R
\name{search_smarts}
\alias{search_smarts}
\title{Search for substances by smarts patterns}
\usage{
search_smarts(smiles, ...)
}
\arguments{
\item{smiles}{a SMILES string}

\item{api_key}{an API_KEY}
}
\description{
For additional optional parameters, see posetra_REST()
}
