% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substance_search.R
\name{search_superstructure}
\alias{search_superstructure}
\title{Search for substances with the query substance as a substructure}
\usage{
search_superstructure(smiles, ...)
}
\arguments{
\item{smiles}{a SMILES string}

\item{api_key}{an API_KEY}
}
\description{
For additional optional parameters, see posetra_REST()
}
