% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substance_search.R
\name{search_synthetic_accessibility_retrosynthesis}
\alias{search_synthetic_accessibility_retrosynthesis}
\title{Search for substances by synthetic accessibility by retrosynthesis}
\usage{
search_synthetic_accessibility_retrosynthesis(smiles, ...)
}
\arguments{
\item{smiles}{a SMILES string}

\item{api_key}{an API_KEY}
}
\description{
For additional optional parameters, see posetra_REST()
}
